+++ async_openai/routes.py
@@ -2,7 +2,7 @@ import aiohttpx
 
 from typing import Optional, Dict, Callable, TYPE_CHECKING
 from async_openai.schemas import *
-from async_openai.utils.config import get_settings, OpenAISettings
+from async_openai.utils.config import get_settings, OpenAISettings, AzureOpenAISettings
 from async_openai.utils.logs import logger
 
 
@@ -38,9 +38,14 @@ class ApiRoutes:
         debug_enabled: Optional[bool] = False,
         on_error: Optional[Callable] = None,
         ignore_errors: Optional[bool] = False,
+        disable_retries: Optional[bool] = None,
+        retry_function: Optional[Callable] = None,
+
         timeout: Optional[int] = None,
         max_retries: Optional[int] = None,
         settings: Optional[OpenAISettings] = None,
+        is_azure: Optional[bool] = None,
+        
         **kwargs
     ):
         self.client = client
@@ -49,11 +54,17 @@ class ApiRoutes:
         self.debug_enabled = debug_enabled
         self.on_error = on_error
         self.ignore_errors = ignore_errors
+        self.disable_retries = disable_retries
+        self.retry_function = retry_function
+
         self.timeout = timeout
         self.max_retries = max_retries
+        self.is_azure = is_azure if is_azure is not None else \
+            isinstance(self.settings, AzureOpenAISettings)
         self.kwargs = kwargs or {}
         self.init_routes()
     
+
 
     def init_routes(self):
         """
@@ -67,9 +78,12 @@ class ApiRoutes:
                     debug_enabled = self.debug_enabled,
                     on_error = self.on_error,
                     ignore_errors = self.ignore_errors,
+                    disable_retries = self.disable_retries,
+                    retry_function = self.retry_function,
                     timeout = self.timeout,
                     max_retries = self.max_retries,
                     settings = self.settings,
+                    is_azure = self.is_azure,
                     **self.kwargs
                 ))
             except Exception as e: